Generic adapter for groups

+++ BAT64-Emulator-JDK/src/main/java/de/treichels/wea/bat64/config/value/ByteValue.java
@@ -8,5 +8,5 @@ public class ByteValue {
 	public final int typeinfo = 52;
 
 	@XmlValue
-	public int value;
+	public short value;
 }
