+++ BAT64-Emulator-JDK/src/main/java/de/treichels/wea/bat64/config/value/TravelValue.java
@@ -13,6 +13,6 @@ public class TravelValue {
 	public final int typeinfo = 45;
 
 	@XmlValue
-	public long value;
+	public short value;
 
 }
